Refactor AppRequest.post
Avoid misunderstanding of using error callback for
succesful POST request.

Due to jQuery design as discussed here:
https://github.com/jquery/jquery/issues/3973

It was considered an error even if the REST API
succesfully created an object but returned empty body.
This is because we tried to force the
`dataType` to `json`,
while in fact PostgREST didn't implement JSON response
body for succesful object creation.

To avoid future misunderstanding, we let jQuery
Ajax determine the `dataType` by omitting the option
for POST request only. This will let us use `success`
callback when the POST request succeeded/return 201.